+++ tests/acceptance/conftest.py
@@ -16,7 +16,7 @@ def pytest_configure(config):
 
 @pytest.fixture
 def cli_runner():
-    return CliRunner(mix_stderr=False)
+    return CliRunner()
 
 
 @pytest.fixture